Potential fix for code scanning alert no. 67: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. This can be achieved by casting one of the operands to uint64 before performing the multiplication. This way, the multiplication will be done using uint64, which has a much larger range and can safely store the result without overflow.

The specific change involves casting n to uint64 before multiplying it by m. This change should be made on line 1906.

+++ libtiff-v4.0.10/libtiff/tif_dirwrite.c
@@ -1903,7 +1903,7 @@ TIFFWriteDirectoryTagTransferfunction(TIFF* tif, uint32* ndir, TIFFDirEntry* dir
 	}
 	if (n==0)
 		n=1;
-	o=_TIFFmalloc((uint32)n * m * sizeof(uint16));
+	o=_TIFFmalloc((uint32)((uint64)n * m) * sizeof(uint16));
 	if (o==NULL)
 	{
 		TIFFErrorExt(tif->tif_clientdata,module,"Out of memory");
